@@ -1,50 +1,38 @@
 {
   "name": "static-cms-next-vercel-template",
-  "version": "1.0.0-beta.0",
-  "license": "MIT",
-  "description": "A starter for a blog powered by Next and Static CMS",
-  "repository": "https://github.com/StaticJsCMS/static-cms-next-vercel-template",
-  "bugs": "https://github.com/StaticJsCMS/static-cms-next-vercel-template/issues",
-  "keywords": [
-    "next",
-    "static",
-    "cms",
-    "vercel"
-  ],
+  "version": "0.1.0",
+  "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "export": "next build && next export -o ./dist",
-    "test": "jest"
+    "lint": "next lint"
   },
   "dependencies": {
     "@staticcms/core": "2.3.3",
-    "crypto": "1.0.1",
-    "date-fns": "2.29.3",
+    "autoprefixer": "10.4.14",
+    "date-fns": "2.30.0",
     "gray-matter": "4.0.3",
     "js-yaml": "4.1.0",
-    "next": "13.1.6",
+    "next": "13.4.2",
     "next-mdx-remote": "4.4.1",
-    "normalize.css": "8.0.1",
+    "postcss": "8.4.23",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-schemaorg": "2.0.0",
     "schema-dts": "1.1.2",
-    "simple-oauth2": "2.5.2",
-    "vercel": "29.3.0"
+    "tailwindcss": "3.3.2"
   },
   "devDependencies": {
-    "@svgr/webpack": "6.5.1",
-    "@types/jest": "29.4.0",
+    "@svgr/webpack": "8.0.1",
     "@types/js-yaml": "4.0.5",
-    "@types/node": "18.14.5",
-    "@types/react": "18.0.28",
-    "cypress": "12.7.0",
-    "eslint-config-next": "13.2.3",
-    "jest": "29.4.3",
+    "@types/node": "20.1.5",
+    "@types/react": "18.2.6",
+    "@types/react-dom": "18.2.4",
+    "cypress": "12.12.0",
+    "eslint": "8.40.0",
+    "eslint-config-next": "13.4.2",
     "js-yaml-loader": "1.2.2",
-    "ts-jest": "29.0.5",
-    "typescript": "4.9.5"
+    "typescript": "5.0.4"
   }
 }
