@@ -1,19 +1,17 @@
 import { format, formatISO } from "date-fns";
 
-type Props = {
+import type { FC } from "react";
+
+export interface DateProps {
   date: Date;
-};
-export default function Date({ date }: Props) {
+}
+
+const Date: FC<DateProps> = ({ date }) => {
   return (
     <time dateTime={formatISO(date)}>
-      <span>{format(date, "LLLL d, yyyy")}</span>
-      <style jsx>
-        {`
-          span {
-            color: #9b9b9b;
-          }
-        `}
-      </style>
+      <span className="text-gray-400">{format(date, "LLLL d, yyyy")}</span>
     </time>
   );
-}
+};
+
+export default Date;
