@@ -1,12 +1,15 @@
 import Head from "next/head";
 import Navigation from "./Navigation";
 
-type Props = {
+import type { FC } from "react";
+
+export interface LayoutProps {
   children: React.ReactNode;
-};
-export default function Layout({ children }: Props) {
+}
+
+const Layout: FC<LayoutProps> = ({ children }) => {
   return (
-    <div className="root">
+    <div className="block px-0 py-16 box-border h-full md:flex md:flex-[1_0_auto]">
       <Head>
         <meta charSet="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
@@ -17,30 +20,9 @@ export default function Layout({ children }: Props) {
       <nav>
         <Navigation />
       </nav>
-      <main>{children}</main>
-      <style jsx>
-        {`
-          .root {
-            display: block;
-            padding: 4rem 0;
-            box-sizing: border-box;
-            height: 100%;
-          }
-          main {
-            display: flex;
-            min-height: 100%;
-          }
-          @media (min-width: 769px) {
-            .root {
-              display: flex;
-              flex: 1 0 auto;
-            }
-            main {
-              flex: 1 0 auto;
-            }
-          }
-        `}
-      </style>
+      <main className="flex min-h-full md:flex-[1_0_auto]">{children}</main>
     </div>
   );
-}
+};
+
+export default Layout;
