@@ -1,28 +1,21 @@
-import { PostContent } from "../lib/posts";
-import Date from "./Date";
-import Link from "next/link";
 import { parseISO } from "date-fns";
+import Link from "next/link";
+import Date from "./Date";
 
-type Props = {
+import type { PostContent } from "@/lib/posts";
+import type { FC } from "react";
+
+export interface PostItemProps {
   post: PostContent;
-};
-export default function PostItem({ post }: Props) {
+}
+
+const PostItem: FC<PostItemProps> = ({ post }) => {
   return (
-    <Link href={"/posts/" + post.slug}>
+    <Link href={"/posts/" + post.slug} className="text-gray-800 inline-block">
       <Date date={parseISO(post.date)} />
-      <h2>{post.title}</h2>
-      <style jsx>
-        {`
-          a {
-            color: #222;
-            display: inline-block;
-          }
-          h2 {
-            margin: 0;
-            font-weight: 500;
-          }
-        `}
-      </style>
+      <h2 className="m-0 font-medium">{post.title}</h2>
     </Link>
   );
-}
+};
+
+export default PostItem;
