@@ -1,56 +1,41 @@
-import React from "react";
+'use client';
+
+import { getAuthor } from "@/lib/authors";
+import { getTag } from "@/lib/tags";
+import { MDXRemote } from "next-mdx-remote";
 import styles from "../../public/styles/content.module.css";
 import Author from "./Author";
 import Copyright from "./Copyright";
 import Date from "./Date";
 import Layout from "./Layout";
+import SocialList from "./SocialList";
+import TagButton from "./TagButton";
 import BasicMeta from "./meta/BasicMeta";
 import JsonLdMeta from "./meta/JsonLdMeta";
 import OpenGraphMeta from "./meta/OpenGraphMeta";
 import TwitterCardMeta from "./meta/TwitterCardMeta";
-import { SocialList } from "./SocialList";
-import TagButton from "./TagButton";
-import { getAuthor } from "../lib/authors";
-import { getTag } from "../lib/tags";
 
-type Props = {
+import type { MDXRemoteSerializeResult } from "next-mdx-remote/dist/types";
+import type { FC } from "react";
+
+export interface PostLayoutProps {
   title: string;
   date: Date;
   slug: string;
   tags: string[];
   author: string;
   description?: string;
-  children: React.ReactNode;
-};
-export default function PostLayout({
-  title,
-  date,
-  slug,
-  author,
-  tags,
-  description = "",
-  children,
-}: Props) {
-  const keywords = tags.map(it => getTag(it)?.name).filter(Boolean);
+  source: MDXRemoteSerializeResult;
+}
+
+const PostLayout: FC<PostLayoutProps> = ({ title, date, slug, author, tags, description = "", source }) => {
+  const keywords = tags.map((it) => getTag(it)?.name).filter(Boolean);
   const authorName = getAuthor(author)?.name;
   return (
     <Layout>
-      <BasicMeta
-        url={`/posts/${slug}`}
-        title={title}
-        keywords={keywords}
-        description={description}
-      />
-      <TwitterCardMeta
-        url={`/posts/${slug}`}
-        title={title}
-        description={description}
-      />
-      <OpenGraphMeta
-        url={`/posts/${slug}`}
-        title={title}
-        description={description}
-      />
+      <BasicMeta url={`/posts/${slug}`} title={title} keywords={keywords} description={description} />
+      <TwitterCardMeta url={`/posts/${slug}`} title={title} description={description} />
+      <OpenGraphMeta url={`/posts/${slug}`} title={title} description={description} />
       <JsonLdMeta
         url={`/posts/${slug}`}
         title={title}
@@ -59,180 +44,37 @@ export default function PostLayout({
         author={authorName}
         description={description}
       />
-      <div className={"container"}>
-        <article>
+      <div className="block max-w-[36rem] w-full my-0 mx-auto py-0 px-6 box-border z-0 md:flex md:flex-col">
+        <article className="flex-[1_0_auto]">
           <header>
-            <h1>{title}</h1>
-            <div className={"metadata"}>
-              <div>
+            <h1 className="mb-2 text-4xl">{title}</h1>
+            <div>
+              <div className="inline-block mr-2">
                 <Date date={date} />
               </div>
-              <div>
+              <div className="inline-block mr-2">
                 <Author author={getAuthor(author)} />
               </div>
             </div>
           </header>
-          <div className={styles.content}>{children}</div>
-          <ul className={"tag-list"}>
+          <div className={styles.content}>{<MDXRemote {...source} />}</div>
+          <ul className="text-right mt-7 p-0">
             {tags.map((it, i) => (
-              <li key={i}>
+              <li key={i} className="inline-block ml-2">
                 <TagButton tag={getTag(it)} />
               </li>
             ))}
           </ul>
         </article>
         <footer>
-          <div className={"social-list"}>
+          <div className="mt-12 text-center">
             <SocialList />
           </div>
           <Copyright />
         </footer>
       </div>
-      <style jsx>
-        {`
-            .container {
-              display: block;
-              max-width: 36rem;
-              width: 100%;
-              margin: 0 auto;
-              padding: 0 1.5rem;
-              box-sizing: border-box;
-              z-index: 0;
-            }
-            .metadata div {
-              display: inline-block;
-              margin-right: 0.5rem;
-            }
-            article {
-              flex: 1 0 auto;
-            }
-            h1 {
-              margin: 0 0 0.5rem;
-              font-size: 2.25rem;
-            }
-            .tag-list {
-              list-style: none;
-              text-align: right;
-              margin: 1.75rem 0 0 0;
-              padding: 0;
-            }
-            .tag-list li {
-              display: inline-block;
-              margin-left: 0.5rem;
-            }
-            .social-list {
-              margin-top: 3rem;
-              text-align: center;
-            }
-
-            @media (min-width: 769px) {
-              .container {
-                display: flex;
-                flex-direction: column;
-              }
-            }
-          `}
-      </style>
-      <style global jsx>
-        {`
-            /* Syntax highlighting */
-            .token.comment,
-            .token.prolog,
-            .token.doctype,
-            .token.cdata,
-            .token.plain-text {
-              color: #6a737d;
-            }
-
-            .token.atrule,
-            .token.attr-value,
-            .token.keyword,
-            .token.operator {
-              color: #d73a49;
-            }
-
-            .token.property,
-            .token.tag,
-            .token.boolean,
-            .token.number,
-            .token.constant,
-            .token.symbol,
-            .token.deleted {
-              color: #22863a;
-            }
-
-            .token.selector,
-            .token.attr-name,
-            .token.string,
-            .token.char,
-            .token.builtin,
-            .token.inserted {
-              color: #032f62;
-            }
-
-            .token.function,
-            .token.class-name {
-              color: #6f42c1;
-            }
-
-            /* language-specific */
-
-            /* JSX */
-            .language-jsx .token.punctuation,
-            .language-jsx .token.tag .token.punctuation,
-            .language-jsx .token.tag .token.script,
-            .language-jsx .token.plain-text {
-              color: #24292e;
-            }
-
-            .language-jsx .token.tag .token.attr-name {
-              color: #6f42c1;
-            }
-
-            .language-jsx .token.tag .token.class-name {
-              color: #005cc5;
-            }
-
-            .language-jsx .token.tag .token.script-punctuation,
-            .language-jsx .token.attr-value .token.punctuation:first-child {
-              color: #d73a49;
-            }
-
-            .language-jsx .token.attr-value {
-              color: #032f62;
-            }
-
-            .language-jsx span[class="comment"] {
-              color: pink;
-            }
-
-            /* HTML */
-            .language-html .token.tag .token.punctuation {
-              color: #24292e;
-            }
-
-            .language-html .token.tag .token.attr-name {
-              color: #6f42c1;
-            }
-
-            .language-html .token.tag .token.attr-value,
-            .language-html
-              .token.tag
-              .token.attr-value
-              .token.punctuation:not(:first-child) {
-              color: #032f62;
-            }
-
-            /* CSS */
-            .language-css .token.selector {
-              color: #6f42c1;
-            }
-
-            .language-css .token.property {
-              color: #005cc5;
-            }
-          `}
-      </style>
     </Layout>
   );
-}
+};
+
+export default PostLayout;
