@@ -1,25 +1,24 @@
-import config from "../../lib/config";
+import config from "@/lib/config";
 import Head from "next/head";
 
-type Props = {
+import type { FC } from "react";
+
+export interface TwitterCardMetaProps {
   url: string;
   title?: string;
   description?: string;
-};
-export default function TwitterCardMeta({ url, title, description }: Props) {
+}
+
+const TwitterCardMeta: FC<TwitterCardMetaProps> = ({ url, title, description }) => {
   return (
     <Head>
       <meta property="twitter:card" content="summary_large_image" />
       <meta property="twitter:site" content={config.twitter_account} />
       <meta property="twitter:url" content={config.base_url + url} />
-      <meta
-        property="twitter:title"
-        content={title ? [title, config.site_title].join(" | ") : ""}
-      />
-      <meta
-        property="twitter:description"
-        content={description ? description : config.site_description}
-      />
+      <meta property="twitter:title" content={title ? [title, config.site_title].join(" | ") : ""} />
+      <meta property="twitter:description" content={description ? description : config.site_description} />
     </Head>
   );
-}
+};
+
+export default TwitterCardMeta;
