@@ -1,10 +1,10 @@
-import authors from "../../meta/authors.yml";
+import authors from "../../content/meta/authors.yml";
 
-export type AuthorContent = {
+export interface AuthorContent {
   readonly slug: string;
   readonly name: string;
   readonly introduction: string;
-};
+}
 
 const authorMap: { [key: string]: AuthorContent } = generateAuthorMap();
 
