@@ -1,4 +1,4 @@
-type Pagination = {
+export interface Pagination {
   page: number | null;
   current: boolean;
   excerpt: boolean;
