@@ -5,13 +5,13 @@ import yaml from "js-yaml";
 
 const postsDirectory = path.join(process.cwd(), "content/posts");
 
-export type PostContent = {
+export interface PostContent {
   readonly date: string;
   readonly title: string;
   readonly slug: string;
   readonly tags?: string[];
   readonly fullPath: string;
-};
+}
 
 let postCache: PostContent[];
 
@@ -39,7 +39,7 @@ export function fetchPostContent(): PostContent[] {
         title: string;
         tags: string[];
         slug: string;
-        fullPath: string,
+        fullPath: string;
       };
       matterData.fullPath = fullPath;
 
@@ -47,9 +47,7 @@ export function fetchPostContent(): PostContent[] {
 
       // Validate slug string
       if (matterData.slug !== slug) {
-        throw new Error(
-          "slug field not match with the path of its content source"
-        );
+        throw new Error("slug field not match with the path of its content source");
       }
 
       return matterData;
@@ -66,16 +64,10 @@ export function fetchPostContent(): PostContent[] {
 }
 
 export function countPosts(tag?: string): number {
-  return fetchPostContent().filter(
-    (it) => !tag || (it.tags && it.tags.includes(tag))
-  ).length;
+  return fetchPostContent().filter((it) => !tag || (it.tags && it.tags.includes(tag))).length;
 }
 
-export function listPostContent(
-  page: number,
-  limit: number,
-  tag?: string
-): PostContent[] {
+export function listPostContent(page: number, limit: number, tag?: string): PostContent[] {
   return fetchPostContent()
     .filter((it) => !tag || (it.tags && it.tags.includes(tag)))
     .slice((page - 1) * limit, page * limit);
